rename: 폴더명 변경 및 위치 수정
